Issue-1126: Slice TaskDetailView more SRP like
