+++ tasks/components/layout/taskDetailView/defaultTaskDetailViewTranslation.ts
@@ -1,0 +1,65 @@
+import type { Languages } from '@helpwave/common/hooks/useLanguage'
+
+export type TaskDetailViewTranslation = {
+  close: string,
+  notes: string,
+  subtasks: string,
+  assignee: string,
+  dueDate: string,
+  status: string,
+  visibility: string,
+  creationTime: string,
+  private: string,
+  public: string,
+  create: string,
+  delete: string,
+  deleteTask: string,
+  deleteTaskDescription: string,
+  publish: string,
+  publishTask: string,
+  publishTaskDescription: string,
+  finish: string
+}
+
+export const defaultTaskDetailViewTranslation: Record<Languages, TaskDetailViewTranslation> = {
+  en: {
+    close: 'Close',
+    notes: 'Notes',
+    subtasks: 'Subtasks',
+    assignee: 'Assignee',
+    dueDate: 'Due date',
+    status: 'Status',
+    visibility: 'Visibility',
+    creationTime: 'Creation time',
+    private: 'private',
+    public: 'public',
+    create: 'Create',
+    delete: 'Delete',
+    deleteTask: 'Delete Task',
+    deleteTaskDescription: 'The Tasks will be irrevocably removed',
+    publish: 'Publish',
+    publishTask: 'Publish task',
+    publishTaskDescription: 'This cannot be undone',
+    finish: 'Finish task',
+  },
+  de: {
+    close: 'Schließen',
+    notes: 'Notizen',
+    subtasks: 'Unteraufgaben',
+    assignee: 'Verantwortlich',
+    dueDate: 'Fälligkeitsdatum',
+    status: 'Status',
+    visibility: 'Sichtbarkeit',
+    creationTime: 'Erstell Zeit',
+    private: 'privat',
+    public: 'öffentlich',
+    create: 'Hinzufügen',
+    delete: 'Löschen',
+    deleteTask: 'Task löschen',
+    deleteTaskDescription: 'Der Tasks wird unwiederruflich gelöscht',
+    publish: 'Veröffentlichen',
+    publishTask: 'Task Veröffentlichen',
+    publishTaskDescription: 'Diese Handlung kann nicht rückgängig gemacht werden',
+    finish: 'Fertigstellen',
+  }
+}
