+++ tasks/components/modals/TaskDetailModal.tsx
@@ -1,7 +1,7 @@
 import { tx } from '@helpwave/common/twind'
-import { Modal, type ModalProps, type ModalHeaderProps } from '@helpwave/common/components/modals/Modal'
-import type { TaskDetailViewProps } from '@/components/layout/TaskDetailView'
-import { TaskDetailView } from '@/components/layout/TaskDetailView'
+import { Modal, type ModalHeaderProps, type ModalProps } from '@helpwave/common/components/modals/Modal'
+import type { TaskDetailViewProps } from '@/components/layout/taskDetailView/TaskDetailView'
+import { TaskDetailView } from '@/components/layout/taskDetailView/TaskDetailView'
 
 export type TaskDetailModalProps =
   Omit<ModalProps, keyof ModalHeaderProps>
@@ -24,7 +24,13 @@ export const TaskDetailModal = ({
       modalClassName={tx(modalClassName)}
       {...modalProps}
     >
-      <TaskDetailView patientId={patientId} wardId={wardId} taskId={taskId} onClose={onClose} initialStatus={initialStatus}/>
+      <TaskDetailView
+        patientId={patientId}
+        wardId={wardId}
+        taskId={taskId}
+        onClose={onClose}
+        initialStatus={initialStatus}
+      />
     </Modal>
   )
 }
